requirements.txt
langchain
streamlit
openai
python-dotenv
langchain-openai
langchain-core
